fix(source): move config to source root
